bio: added affine gap cost matrix
Added a affine gap cost matrix that demonstrates how to build more advanced
cost matrices. The affine gap cost matrix forces a higher penalty for starting
a gap, but less of a penalty for extending a gap.

Added an "Alignment" abstraction that builds an object that encapsulates an
alignment. This should make it easier to build useful applications that use
this library (before all you could do is print the cigar form of the two
sequences).